[#16] Settings activity exists, renders the wallpaper behind it
Lots of updates here, overly large commit unfortunately. This adds some
user interface, refactored the API a bit to support on demand wallpaper
generation, included a broadcast for active wallpaper completion as
well so that on first launch, the activity can wait for the initial
task to complete.
