+++ LiveWallpaperQuotes/app/src/main/java/com/stanleyidesis/livewallpaperquotes/ui/activity/LWQSettingsActivity.java
@@ -1,23 +1,178 @@
 package com.stanleyidesis.livewallpaperquotes.ui.activity;
 
-import android.app.Activity;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
+import android.os.Build;
 import android.os.Bundle;
+import android.support.design.widget.TabLayout;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+import android.support.v4.view.ViewPager;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.graphics.Palette;
+import android.support.v7.widget.Toolbar;
+import android.util.Log;
+import android.view.SurfaceHolder;
+import android.view.SurfaceView;
+import android.view.View;
 
+import com.stanleyidesis.livewallpaperquotes.LWQApplication;
+import com.stanleyidesis.livewallpaperquotes.LWQPreferences;
 import com.stanleyidesis.livewallpaperquotes.R;
+import com.stanleyidesis.livewallpaperquotes.api.Callback;
+import com.stanleyidesis.livewallpaperquotes.api.LWQDrawScript;
 
 /**
  * Created by stanleyidesis on 7/11/15.
  */
-public class LWQSettingsActivity extends Activity {
+public class LWQSettingsActivity extends AppCompatActivity implements SurfaceHolder.Callback {
+
+    LWQDrawScript drawScript;
+    BroadcastReceiver newWallpaperBroadcastReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            if (drawScript != null) {
+                drawScript.draw();
+            }
+        }
+    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         setContentView(R.layout.activity_lwq_settings);
+
+        SurfaceView surfaceView = (SurfaceView) findViewById(R.id.surface_lwq_settings);
+        surfaceView.getHolder().addCallback(this);
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar_lwq_settings);
+        setSupportActionBar(toolbar);
+
+        ViewPager viewPager = (ViewPager) findViewById(R.id.viewpager_lwq_settings);
+        viewPager.setAdapter(new PagerAdapter(getSupportFragmentManager()));
+
+        TabLayout tabLayout = (TabLayout) findViewById(R.id.tab_layout_lwq_settings);
+//        tabLayout.setupWithViewPager(viewPager);
     }
 
     @Override
     protected void onResume() {
         super.onResume();
+        toggleHideyBar();
+        IntentFilter intentFilter = new IntentFilter(getString(R.string.broadcast_new_wallpaper_available));
+        registerReceiver(newWallpaperBroadcastReceiver, intentFilter);
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        unregisterReceiver(newWallpaperBroadcastReceiver);
+    }
+
+    @Override
+    public void surfaceCreated(SurfaceHolder holder) {
+        drawScript = new LWQDrawScript(new Palette.PaletteAsyncListener() {
+            @Override
+            public void onGenerated(Palette palette) {
+                drawScript.draw();
+            }
+        }, holder);
+        if (LWQPreferences.isFirstLaunch()) {
+            return;
+        }
+        LWQApplication.getWallpaperController().retrieveActiveWallpaper(new Callback<Boolean>() {
+            @Override
+            public void onSuccess(Boolean aBoolean) {
+                if (drawScript != null) {
+                    drawScript.draw();
+                }
+            }
+
+            @Override
+            public void onError(String errorMessage) {
+
+            }
+        }, true);
+    }
+
+    @Override
+    public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
+        if (drawScript != null) {
+            drawScript.setSurfaceHolder(holder);
+            drawScript.draw();
+        }
+    }
+
+    @Override
+    public void surfaceDestroyed(SurfaceHolder holder) {
+
+    }
+
+    /**
+     * Detects and toggles immersive mode (also known as "hidey bar" mode).
+     */
+    void toggleHideyBar() {
+        /*if (true) {
+            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
+                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
+            return;
+        }*/
+
+        // The UI options currently enabled are represented by a bitfield.
+        // getSystemUiVisibility() gives us that bitfield.
+        int uiOptions = getWindow().getDecorView().getSystemUiVisibility();
+        int newUiOptions = uiOptions;
+        boolean isImmersiveModeEnabled =
+                ((uiOptions | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY) == uiOptions);
+        if (isImmersiveModeEnabled) {
+            Log.i(getClass().getSimpleName(), "Turning immersive mode mode off. ");
+        } else {
+            Log.i(getClass().getSimpleName(), "Turning immersive mode mode on.");
+        }
+
+        // Navigation bar hiding:  Backwards compatible to ICS.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.ICE_CREAM_SANDWICH) {
+            newUiOptions ^= View.SYSTEM_UI_FLAG_HIDE_NAVIGATION;
+        }
+
+        // Status bar hiding: Backwards compatible to Jellybean
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+            newUiOptions ^= View.SYSTEM_UI_FLAG_FULLSCREEN;
+        }
+
+        // Immersive mode: Backward compatible to KitKat.
+        // Note that this flag doesn't do anything by itself, it only augments the behavior
+        // of HIDE_NAVIGATION and FLAG_FULLSCREEN.  For the purposes of this sample
+        // all three flags are being toggled together.
+        // Note that there are two immersive mode UI flags, one of which is referred to as "sticky".
+        // Sticky immersive mode differs in that it makes the navigation and status bars
+        // semi-transparent, and the UI flag does not get cleared when the user interacts with
+        // the screen.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
+            newUiOptions ^= View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY;
+        }
+
+        getWindow().getDecorView().setSystemUiVisibility(newUiOptions);
+    }
+
+    class PagerAdapter extends FragmentPagerAdapter {
+
+        public PagerAdapter(FragmentManager fm) {
+            super(fm);
+        }
+
+        @Override
+        public Fragment getItem(int position) {
+            return null;
+        }
+
+        @Override
+        public int getCount() {
+            return 0;
+        }
     }
 }
