+++ LiveWallpaperQuotes/app/src/main/res/values/styles.xml
@@ -1,8 +1,10 @@
 <resources>
 
     <!-- Base application theme. -->
-    <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
+    <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
+        <item name="colorPrimary">#FFFFFF</item>
+        <item name="colorPrimaryDark">#DDDDDD</item>
+        <item name="colorAccent">#AAAAAA</item>
     </style>
 
 </resources>
